This should now be working.  Analyze particle is disabled by default when a collection is selected.  When an "Atom" is selected, the analyze particle button is enabled if the collection is an ATOFMS or AMS collection.  Double clicking on an atom is likewise enabled if the collection is an ATOFMS or AMS collection.
This was accomplished by changing lines 1346-1369.  Previously the check inside ListSelectionListener merely checked to see that an Atom had already been selected before enabling the particle analyze button.  I modified the check to include a check for the datatype.
I also encapsulated the MouseListener in a similar check.  I inserted a line to ensure that the analyze particle button is disabled by default when a collection is selected.