Changed the ReadSpec method to take in a date object as well as a filename to fix the timezone bug.
ATOFMSDataSetImporter now looks up the date in the .set file and passes it in to readSpec (lines 475-80,493).

Changed the experiments so that they work with the new ReadSpec constructor by adding date objects.

Changed  ParticleAnalyzeWindow to grab the time string from the same row it gets the filename from and converts it into a date object so that ReadSpec can be called properly.

Emma Turetsky

Reviewed by Sami Benzaid